+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sikka/hawa
 
+## 0.19.39
+
+### Patch Changes
+
+- - Added background to `Radio`
+  - Fixed responsiveness of `PricingPlans`
+  - Added priceless option to `PricingCard`
+
 ## 0.19.38
 
 ### Patch Changes